Test system manager deployment triggered by Ansible
