+++ ansible/tasks/setup-system-manager.yml
@@ -1,0 +1,7 @@
+---
+- name: Deploy system manager
+  ansible.builtin.shell: |
+    . /nix/var/nix/profiles/default/etc/profile.d/nix-daemon.sh
+    cd /tmp
+    nix run /flake#system-manager -- switch --flake /flake
+  become: true
